backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic[email]>=2.6
pydantic-settings>=2.1
python-jose[cryptography]>=3.3
bcrypt>=4.1
google-cloud-secret-manager>=2.18
httpx>=0.27
python-multipart>=0.0.6
asyncssh>=2.14
# Phase 5: Data Management + Visualization
google-cloud-storage>=2.14
pdfplumber>=0.11
matplotlib>=3.8
anndata>=0.10
scanpy>=1.10
meilisearch>=0.31
kubernetes>=29.0
# Phase 6: GitOps + Package Management
PyGithub>=2.1.0
pyyaml>=6.0
